Move the Model.reopen from relationships/ext.js into model/model to make it easier to import model
