+++ addon/relationships.js
@@ -5,8 +5,6 @@
 import belongsTo from "ember-data/-private/system/relationships/belongs-to";
 import hasMany from "ember-data/-private/system/relationships/has-many";
 
-import "ember-data/-private/system/relationships/ext";
-
 export {
   belongsTo,
   hasMany
